Renamed the controller
